start refactoring hooks
