+++ internal/orchestrator/tasks/hookvars.go
@@ -1,4 +1,4 @@
-package hook
+package tasks
 
 import (
 	"bytes"
