+++ internal/orchestrator/tasks/task.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	v1 "github.com/garethgeorge/backrest/gen/go/v1"
-	"github.com/garethgeorge/backrest/internal/hook"
 	"github.com/garethgeorge/backrest/internal/oplog"
 	"github.com/garethgeorge/backrest/internal/orchestrator/repo"
 	"go.uber.org/zap"
@@ -35,7 +34,7 @@ type TaskRunner interface {
 	// UpdateOperation updates the operation in storage. It must be called after CreateOperation.
 	UpdateOperation(*v1.Operation) error
 	// ExecuteHooks
-	ExecuteHooks(events []v1.Hook_Condition, vars hook.HookVars) error
+	ExecuteHooks(events []v1.Hook_Condition, vars HookVars) error
 	// OpLog returns the oplog for the operations.
 	OpLog() *oplog.OpLog
 	// GetRepo returns the repo with the given ID.
