+++ internal/orchestrator/tasks/taskstats.go
@@ -6,7 +6,6 @@ import (
 	"time"
 
 	v1 "github.com/garethgeorge/backrest/gen/go/v1"
-	"github.com/garethgeorge/backrest/internal/hook"
 	"github.com/garethgeorge/backrest/internal/oplog"
 	"github.com/garethgeorge/backrest/internal/oplog/indexutil"
 )
@@ -72,7 +71,7 @@ func (t *StatsTask) Run(ctx context.Context, st ScheduledTask, runner TaskRunner
 	if err := statsHelper(ctx, st, runner); err != nil {
 		runner.ExecuteHooks([]v1.Hook_Condition{
 			v1.Hook_CONDITION_ANY_ERROR,
-		}, hook.HookVars{
+		}, HookVars{
 			Task:  st.Task.Name(),
 			Error: err.Error(),
 		})
